finished carousel and added alerts

+++ src/components/UserGallery/MyFeaturedImage.js
@@ -3,6 +3,7 @@ import {useParams, useHistory} from 'react-router-dom';
 import { useSelector, useDispatch } from 'react-redux';
 import ColoredInImage from '../Images/ColoredInImage';
 import {saveSvgAsPng} from 'save-svg-as-png';
+import swal from 'sweetalert';
 
 
 
@@ -31,18 +32,30 @@ export default function MyFeaturedImage(props) {
     }
 
     const deleteImage = () => {
-        fetch(`http://localhost:3000/user_images/${id}`,{
-            method: 'DELETE',
-            headers: {'Authorization': `bearer ${token}`}
+        swal({
+            title: "Are you sure?",
+            text: "Once deleted, you will not be able to recover this image",
+            icon: "warning",
+            buttons: true,
+            dangerMode: true,
         })
-        .then(r => r.json())
-        .then(data => {
-            dispatch({
-                type: 'SET_IMAGES',
-                payload: data
-              });
-              history.push(`/mygallery`)
-            })      
+        .then((willDelete) => {
+            if (willDelete) {
+                fetch(`http://localhost:3000/user_images/${id}`,{
+                    method: 'DELETE',
+                    headers: {'Authorization': `bearer ${token}`}
+                })
+                .then(r => r.json())
+                .then(data => {
+                    dispatch({
+                        type: 'SET_IMAGES',
+                        payload: data
+                      });
+                      swal("Image Deleted", "This image has been removed from your gallery", "success");
+                      history.push(`/mygallery`)
+                })        
+            }
+          });
     }
 
     const editImage = () => {
@@ -55,6 +68,7 @@ export default function MyFeaturedImage(props) {
 
     const saveImage = () => {
         saveSvgAsPng(document.getElementById("svg-image"), "myimage.png");
+        swal("Download Complete", "This image has been saved to your Downloads", "success");
     }
 
     return(
